add swagger

+++ internal/api/points.go
@@ -44,7 +44,7 @@ func (p *PointsHandler) Points(c *gin.Context) {
 		fmt.Println("Error converting string to int:", err)
 		return
 	}
-	res := p.pointsService.Points(userIDInt, chainType)
+	res := p.pointsService.Points(uint(userIDInt), chainType)
 	c.JSON(res.Code, res)
 }
 
@@ -91,6 +91,19 @@ func (p *PointsHandler) PointsDetail(c *gin.Context) {
 	c.JSON(res.Code, res)
 }
 
+// PointsDetail 获取用户邀请明细
+// @Summary 获取用户邀请明细
+// @Description 根据链类型和用户 ID 获取用户的积分明细数据。支持的链类型：sol（Solana）、eth（Ethereum）、bsc（Binance Smart Chain）。
+// @Tags 用户
+// @Accept json
+// @Produce json
+// @Security ApiKeyAuth
+// @Param chain_type path string true "链类型（sol、eth、bsc）"
+// @Param page query string true "分页页码"
+// @Param limit query string true "每页数量"
+// @Success 200 {object} response.Response{data=response.PointsDetailsResponse} "成功返回用户积分明细"
+// @Failure 500 {object} response.Response "服务器内部错误"
+// @Router /points/{chain_type}/detail [get]
 func (p *PointsHandler) InvitedPointsDetail(c *gin.Context) {
 	userID, errResp := GetUserIDFromContext(c)
 	if errResp != nil {
